requirements.txt
opencv-python==4.6.0.66
imutils==0.5.4
Pillow==9.2.0
numpy==1.21.6